Remove base application class
basic_application_t and application classes are removed. Before basic_application maped events in handler factories, but now worker does it. Worker is renamed to dispatcher, because dispatching of events is his main functionality for user.
Now user must provide for function 'run' class that has method 'init'. This method must take a pointer on dispatcher_t as argument. This is minimal requirement to create an application.