Make service_manager more rubust
Now service_manager initializes (with resolver and logger) even if runtime with locator is down. If a service is disconnected then it tries to connect on every call. Call to service no longer throws, just returns future with error. So user of the framework can just do calls and must not think about reconnects.