Remove session when an error from a service is received.
Error message always finishes a session, so it's not necessary to wait choke to release resources.
Also this commit fixes a bug. If an error occures twice (error from service and then losing of the connection) then corresponding exceptions are passed to a handler twice too. In this case the handler throws exception "future already satisfied" that leads to crashing of a program.