[Detail] Allow to invoke `connect` multiple times.
If the service is disconnected, a new connection should be established.
If the service is in connecting state - a new handle (future) is created, but there are no other connection attempts performed.
At last if the service is already connected the newly created promise is immediately set to be ready.