[Bug Fix] Fixed dangling reference.
According to the MessagePack documentation, zone class owns almost all the data it represents except raw type, which is pointed and it's the user responsibility to manage its lifetime until there are no msgpack::object objects in use.
This behavior forces me to perform copy operation on unpacked buffer, which is highly ineffective and costly.

I need further research to deal with it.